+++ examples/publisher/source/ts/configuration.ts
@@ -355,6 +355,24 @@ export const adConfiguration: Moli.MoliConfig = {
       ]
     },
 
+    /* manualy slots */
+    {
+      domId: 'manual-adslot',
+      position: 'in-page',
+      behaviour: {
+        loaded: 'manual'
+      },
+      adUnitPath: '/55155651/prebid_test',
+      sizes: [ [ 300, 250 ] ],
+      passbackSupport: true,
+      sizeConfig: [
+        {
+          mediaQuery: '(min-width: 0px)',
+          sizesSupported: [ [ 300, 250 ] ]
+        }
+      ]
+    },
+
   ],
   // -----------------------------
   // ----- standard config -------