+++ examples/single-page-app/index.ts
@@ -7,10 +7,6 @@ import prebid from 'prebid.js';
 import 'prebid.js/modules/consentManagement';
 import 'prebid.js/modules/currency';
 import 'prebid.js/modules/appnexusBidAdapter';
-import 'prebid.js/modules/pubmaticBidAdapter';
-import 'prebid.js/modules/smartadserverBidAdapter';
-import 'prebid.js/modules/teadsBidAdapter';
-import 'prebid.js/modules/unrulyBidAdapter';
 
 import { initAdTag } from '@highfivve/ad-tag';
 import { adConfiguration } from './source/ts/configuration';